+++ app/src/main/java/org/hyperledger/besu/controller/QbftBesuControllerBuilder.java
@@ -282,7 +282,10 @@ public class QbftBesuControllerBuilder extends BesuControllerBuilder {
     final MessageTracker duplicateMessageTracker =
         new MessageTracker(qbftConfig.getDuplicateMessageLimit());
 
-    final MessageFactory messageFactory = new MessageFactory(nodeKey, blockEncoder);
+    final MessageFactory messageFactory =
+        isLegacyBftProtocolEncodingEnabled
+            ? MessageFactory.withLegacyEncoding(nodeKey, blockEncoder)
+            : new MessageFactory(nodeKey, blockEncoder);
 
     QbftRoundFactory qbftRoundFactory =
         new QbftRoundFactory(